+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [15.1.0] (melonJS 2) - _2023-xx-xx_
 
+### Added
+- Renderer: add a new `convertToBlob()` method to `CanvasTexture` 
+
 ### Fixed
 - Renderer: add missing export for the `CanvasTexture` class